fixed: basic clean up (signed/unsigned mixups, printfs formats)
git-svn-id: https://opensg.vrsource.org/svn/trunk@2060 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/Base/Base/OSGMultiPool.h
@@ -106,7 +106,7 @@ class MultiPool
     {
       public:
       
-        static const int BLOCKSIZE = 16; //!< number of elements per block
+        static const UInt16 BLOCKSIZE = 16; //!< number of elements per block
         
         TypeStore(int size);
